Chats client code refactor
